robot/requirements.txt
# Rovy Raspberry Pi Client Requirements
# Controls rover hardware and streams to PC server

# WebSocket client
websockets>=11.0

# Serial communication with rover
pyserial>=3.5

# Camera
opencv-python>=4.5.0

# Audio (ReSpeaker)
pyaudio>=0.2.11
numpy>=1.21.0

# For local audio playback
sounddevice>=0.4.0
soundfile>=0.12.0

